Refactor storage.cassandra package.
